bash_aliases: aliases for nmap now ask for sudo password if the scan requires root
If something requires root, alias will run "sudo command" instead of that command.
Running only command will only say "This must be run as root." But there is one
expection. Nmap-traceroute doesn't require root, but it will say "traceroute requires
root, cancelling" and tells that the host is up (or down depending on it). I added sudo
for it, because we usually want to see traceroute if we want traceroute.

This is the longest commit message which I have ever written. Maybe I will write
longer someday...