systemd: snapd.refresh.service override
It fails on laptops thanks to not being able to do DNS resolution thanks
to network connection not existing during boot.

Now it fails to `Download snap "ubuntu-core" (423) from channel "stable"
(cannot authenticate to snap store: Provided email/password is not
correct.)` which is process and appears to not be my issue.