etc/apt/sources.list: enable Debian debugsym repos
It seems that I am always going to enable it sooner or later anyway, so
why woulnd't I have it enabled for quick installing when I do need it?

Example: KDE Connect crashed on login, and asked me to report it, but
the reporter app warned that there is no address to report it and debug
information had one or two stars and said that it's likely bad quality
and I think this is due to missing debug symbols which I then installed.

Naturally after installing them, I am unable to reproduce the issue, but
that is beside the point.