i3: prepare for j4-dmenu-desktop daemon-like mode & float Firefox
When I login and attempt to use dmenu, it will take a moment before that
is possible and I thought that having it load before I try to use it
would be nice. Then I found an issue about it and PR that closed it, but
didn't notice the commit is not in any released version so I typed this
in advance.

Firefox is going to annoy me by not floating sooner or later, so I think
it's better to float it and keep using it as the only thing in one
workspace.