i3: move to xfce4-terminal & introduce $MyPreferredTerminal
xfce4-terminal does the same things I had urxvt do and works with emojis
that were my problem with urxvt and even does copy-pasting with
Ctrl-Shift-C/V which terminology for some reason failed to do and is
probably even lighter.

I also disliked the idea of having to change the story of what is my
preferred terminal emulator everywhere, so now there is a single
variable which I can throw around as I need.