i3: stack by default, centerize titles & add compton comment
I seem to be unable to find happiness with default and often after
opening multiple windows my first thing to do is moving to stacked or
tabbed layout. Stacked layout seems easier to split than windowed which
is still easy to reach and centerized titles probably benefit windowed
layout the most.

The compton example makes everything transparent and is horrible.