systemd: copy matterbridge restarter into gitea one
It used to stop working randomly without a good reason, but that is
likely fixed upstream a long time ago and while I removed it from cron,
these units exist so should the issue recur, I can throw these back in.

The cron wasn't randomized though, but I don't think there is harm in
this being a bit random.